[test] : 인가 관련 API 컨트롤러 위치 이동으로 인한 테스트 코드 위치 변경
